Initial Commit: Designed loaders for modules, datasets and configuration.
This commit includes the skeletons for loaders that can load dataset
information and configuration information, as well as the API for
modules in the LGP system with a loader for these modules.

Two main implementations of these APIs were implemented:

    - CsvDatasetLoader: for loading a dataset from a CSV file.
    - JsonConfigLoader: for loading configuration information from a
      JSON file.

The skeletons of these loaders will be used by the environment to gather
the dependencies that it needs in order to perform the core evolution
process.