Add skeleton for instructions, registers, environment.
Instructions will primarily consists of an Operation and operate on
Arguments (which themself come from Registers). The basic outlines for
these components are now available and will be used in further commits
to define what Instructions in the LGP system look like.

Documentation is provided with the new components to provide an
understanding of how they are used.

The Environment has also been further implemented to take various
loaders in its constructors which are used to build an environment for
the LGP system to use in the evolution process. This environment
includes elements such as:
    - Data set
    - Constants
    - Operations
    - Configuration

This will be built upon in further commits.