Further implementation of core evolution constructs.
Changes made to handle generating programs, including implementation of
an abstract program generator.

Notably, Instructions were changed such that they do not depend on a
register set being given at execution time - instead the instruction
generator has access to everything within the LGP environment and is
responsible for marshalling the dependencies to the instruction
implementation.

This idea is used in the program generator as well, leaving the
dependency resolution to be done by any implementor.

In further commits, examples of implementations will be given that give
an indiciation of how this process works.