Add ability to specify class/input attributes from dataset in config.
The configuration now allows for the class attribute of instances in a
dataset to be specified through the `classAttributeIndex` field. This
field expects an integer that refers to an attribute by index from a set
of instances.

Furthermore, the indices of attributes in an instance that are to be
regarded as "inputs" can be specified through configuration fields (that
is, attributes which should be loaded into the register set as inputs).

The input attributes are specified using a range between two fields:
`inputAttributesLowIndex` the lower bound, and
`inputAttributesHighIndex` the upper bound. The interval
`inputAttributesLowIndex..inputAttributesHighIndex` will be generated
and used to determine which attributes are to be used for input.

By default, values of 0 are used for all indices, so it will need to be
advised in the documentation to set values for these fields or make sure
users are aware of the defaults.