Add revised and stable version of search operators.
This commit addresses some of the issues mentioned in
a84b9c6ee6461f7229282c0e23822ce9b1b1ea2b, and should have a stable and
complete version of the search operators (particularly mutation, as it
was giving some issues).

I've tested the implementation on a simple benchmark problem (which I
will include as an example in a futher commit), and the results are as
expected (with a perfect individual being found some of the time).

The architectural changes were mainly made in previous commits, and
involved making the Instruction and Program modules a bit more
restricted in their interface. The library implementations of these will
be introduced in further commits.