Refactor module loading for registered components.
This commit makes the loading of modules used as registered components
safer and more modular.

In terms of safety, the loading no longer requires unchecked casts.
Reified generics are used to ensure that the type the module is to be
casted is a valid cast for that module.

In the cases where the cast is not valid a custom exception is given
that provides information why this may have happened.

Furthemore, the registered module type interface has been given so that
different enumerations of registered module types can be provided based
on how the environment is used.

The core registered module type enumeration provides values for all of
the core modules that are needed by the environment, but custom logic
may need other modules (or allow further core modules to be resolved at
registered component time).