Add class to allow easier problem definition.
Now, instead of defining the problem directly in the main method of some
class, we have a `Problem` class which can be "filled in" as necessary
to define a problem.

This problem can then be created in a main method and solved using any
method defined by the problem. This will hopefully allow more
flexibility, as well as giving a clearer idea of how the different parts
are pieced together in order to solve a problem.