Modularise fitness context and evaluation.
Previously, the way in which a fitness context gathered program outputs
for use in fitness function application was static --- limiting the
extensibility of the system to single-output programs only.

This commit introduces a modularised version of the `FitnessContext` and
`FitnessEvaluator` classes that can handle further methods of program
output aggregation (e.g. multiple-output, weighted-output). By default,
the implementation `SingleOutputFitnessContext` provides the same
behaviour as before these changes were made.

Of note with this change is that the particular `FitnessContext`
implementation will now need to be registered with the environment:
the `CoreModuleType.FitnessContext` provides the key to which the
implementation should be registered. Examples of this will be given in
the `lgp.examples` package in further commits.

This commit closes issue #5.