Add ability to seed RNG singleton; Add copy method.
Beforehand, an environment had no control over the RNG, but this commit
allows the RNG to be a central singleton that is accessed through the
environment. This is the beginning of allowing modules to gather random
numbers through the environment, providing the ability for deterministic
runs (by seeding the RNG).

This did require some fairly significant changes -- namely:

    - The module container now needs to have a reference to the
      environment it belongs to. This allows the modules to have their
      environment reference updated when the environment is copied.
    - The environment constructor has a new parameter
      (`randomStateSeed`). This parameter is optional so it shouldn't
      break any existing setups.

Further commits will use this new logic to facilitate the ability to
produce determinstic runs of the system.