Update core components to use central RNG.
The core modules now utilise the RNG provided by the environment rather
than constructing a new instance each time an RNG is required. Not only
will this improve memory consumption and performance (i.e. not
allocating as many objects); but, it allows for the ability to produce
deterministic runs through the use of a fixed seed.

Furthermore, a deep copy method was added to the `EvolutionModel` class
which will the copy of a model to have a copy of its environment too.