Add ability for Java callers to access registered modules.
Previously, Java callers were unable to request registered modules as
the API exposed an inlined function which used reified generics -- one
which could not be used from Java.

To improve the Java interoperability, functions have been added to
provide the same functionality to Java users. There is a caveat however:
due to Java not having reified generics, the same guarantees of type
safety can't be made and there may be cases where a `ClassCastException`
is given. There is not much that can be done about this, other than
hoping any users will understand what they are doing.