Add validation for configuration information.
Adds a method to the configuration class to ensure that the
configuration is valid. In some situations, the state of configuration
could cause the system to enter a deadlock. This fix resolves the issue.

Additionally, the class has been renamed to make things clearer -- as
well as consumers of the class being updated.