First cut - async sequential training.
This commit introduces a basic async sequential training implementation.
Essentially, there are two threads (trainer and requester) which
communicate via an actor.

The trainer will send progress updates to the actor which will update
the shared progress. The requester can request the progress and the
actor will send it back via a channel.

Further commits will simplify this process and allow the requester to
subscribe to progress updates and execute a callback.