Modified async training implementation to use a BroadcastChannel.
Using a BroadcastChannel offers a better API for consumers of the async
training functionality. Basically, the concepts are as follows:

1. Begin training which will create a training coroutine and return an
object that represents the training
2. Optionally subscribe to progress updates from the training coroutine.
The training coroutine will yield to the susbcription coroutines
allowing them to get the current progress
3. Wait for the training coroutine to complete

Soon, I will implement similar functionality for the DistributedTrainer
but it will require some more thought to ensure that the different
threads co-ordinate their communication in a sane way.