Initial implementation for multiple output programs.
Single output programs are still behaving as expected. Need to test with
a multiple output dataset to see what else needs to change for it to
work.

At the moment I'm a bit unhappy with the fitness function provider
addition as there is some casting involved that should ideally be
removed, but it seems that type erasure is causing some problems.

The Java example is still broken and has been commented out for now, but
will need to address that in further commits.